Merge master into dev/cms_unix

+++ build.sh
@@ -6,7 +6,7 @@ usage()
     echo "managed - optional argument to build the managed code"
     echo "native - optional argument to build the native code"
     echo "The following arguments affect native builds only:"
-    echo "BuildArch can be: x64, x86, arm, arm64"
+    echo "BuildArch can be: x64, x86, arm, arm-softfp, arm64"
     echo "BuildType can be: debug, release"
     echo "clean - optional argument to force a clean build."
     echo "verbose - optional argument to enable verbose build output."
@@ -87,11 +87,13 @@ prepare_native_build()
 
     # Generate version.c if specified, else have an empty one.
     __versionSourceFile=$__scriptpath/bin/obj/version.c
-    if [ $__generateversionsource == true ]; then
-        $__scriptpath/Tools/corerun $__scriptpath/Tools/MSBuild.exe "$__scriptpath/build.proj" /t:GenerateVersionSourceFile /p:NativeVersionSourceFile=$__scriptpath/bin/obj/version.c /p:GenerateVersionSourceFile=true /v:minimal $__OfficialBuildIdArg
-    else
-        __versionSourceLine="static char sccsid[] __attribute__((used)) = \"@(#)No version information produced\";"
-        echo $__versionSourceLine > $__versionSourceFile
+    if [ ! -e "${__versionSourceFile}" ]; then
+        if [ $__generateversionsource == true ]; then
+            $__scriptpath/Tools/dotnetcli/dotnet $__scriptpath/Tools/MSBuild.exe "$__scriptpath/build.proj" /t:GenerateVersionSourceFile /p:GenerateVersionSourceFile=true /v:minimal
+        else
+            __versionSourceLine="static char sccsid[] __attribute__((used)) = \"@(#)No version information produced\";"
+            echo $__versionSourceLine > $__versionSourceFile
+        fi
     fi
 }
 
@@ -102,7 +104,7 @@ build_managed()
     __binclashlog=$__scriptpath/binclash.log
     __binclashloggerdll=$__scriptpath/Tools/Microsoft.DotNet.Build.Tasks.dll
 
-    $__scriptpath/Tools/corerun $__scriptpath/Tools/MSBuild.exe "$__buildproj" /m /nologo /verbosity:minimal "/fileloggerparameters:Verbosity=normal;LogFile=$__buildlog" "/l:BinClashLogger,$__binclashloggerdll;LogFile=$__binclashlog" /t:Build /p:ConfigurationGroup=$__BuildType /p:TargetOS=$__BuildOS /p:OSGroup=$__BuildOS /p:SkipTests=$__SkipTests /p:COMPUTERNAME=$(hostname) /p:USERNAME=$(id -un) /p:TestNugetRuntimeId=$__TestNugetRuntimeId $__UnprocessedBuildArgs
+    $__scriptpath/Tools/dotnetcli/dotnet $__scriptpath/Tools/MSBuild.exe "$__buildproj" /m /nologo /verbosity:minimal "/flp:Verbosity=normal;LogFile=$__buildlog" "/flp2:warningsonly;logfile=$__scriptpath/msbuild.wrn" "/flp3:errorsonly;logfile=$__scriptpath/msbuild.err" "/l:BinClashLogger,$__binclashloggerdll;LogFile=$__binclashlog" /p:ConfigurationGroup=$__BuildType /p:TargetOS=$__BuildOS /p:OSGroup=$__BuildOS /p:SkipTests=$__SkipTests /p:COMPUTERNAME=$(hostname) /p:USERNAME=$(id -un) /p:TestNugetRuntimeId=$__TestNugetRuntimeId $__UnprocessedBuildArgs
     BUILDERRORLEVEL=$?
 
     echo
@@ -160,14 +162,10 @@ __nativeroot=$__scriptpath/src/Native
 __packageroot=$__scriptpath/packages
 __sourceroot=$__scriptpath/src
 __rootbinpath="$__scriptpath/bin"
-__msbuildpackageid="Microsoft.Build.Mono.Debug"
 __generateversionsource=false
-__msbuildpackageversion="14.1.0.0-prerelease"
-__msbuildpath=$__packageroot/$__msbuildpackageid.$__msbuildpackageversion/lib/MSBuild.exe
 __buildmanaged=false
 __buildnative=false
 __TestNugetRuntimeId=win7-x64
-__OfficialBuildIdArg=
 
 # Use uname to determine what the CPU is.
 CPUName=$(uname -p)
@@ -215,22 +213,12 @@ case $OSName in
 
     Linux)
         __HostOS=Linux
-        source /etc/os-release
-        if [ "$ID" == "centos" ]; then
-            __TestNugetRuntimeId=centos.7-x64
-        elif [ "$ID" == "rhel" ]; then
-            __TestNugetRuntimeId=rhel.7-x64
-        elif [ "$ID" == "ubuntu" ]; then
-            if [ $VERSION_ID == "16.04" ]; then
-                __TestNugetRuntimeId=ubuntu.16.04-x64
-            else
-                __TestNugetRuntimeId=ubuntu.14.04-x64
-            fi
-        elif [ "$ID" == "debian" ]; then
-            __TestNugetRuntimeId=debian.8-x64
-        else
-            echo "Unsupported Linux distribution '$ID' detected. Configuring as if for Ubuntu."
+        if [ ! -e /etc/os-release ]; then
+            echo "Cannot determine Linux distribution, assuming Ubuntu 14.04"
             __TestNugetRuntimeId=ubuntu.14.04-x64
+        else
+            source /etc/os-release
+            __TestNugetRuntimeId=$ID.$VERSION_ID-$__BuildArch
         fi
         ;;
 
@@ -289,6 +277,9 @@ while :; do
         arm)
             __BuildArch=arm
             ;;
+        arm-softfp)
+            __BuildArch=arm-softfp
+            ;;
         arm64)
             __BuildArch=arm64
             ;;
@@ -319,6 +310,10 @@ while :; do
         clang3.7)
             __ClangMajorVersion=3
             __ClangMinorVersion=7
+            ;;
+        clang3.8)
+            __ClangMajorVersion=3
+            __ClangMinorVersion=8
             ;;
         freebsd)
             __BuildOS=FreeBSD
@@ -363,11 +358,7 @@ while :; do
             __ServerGC=1
             ;;
         *)
-          if [[ $1 == "/p:OfficialBuildId="* ]]; then
-            __OfficialBuildIdArg=$1
-          else
             __UnprocessedBuildArgs="$__UnprocessedBuildArgs $1"
-          fi
     esac
 
     shift
